Conflicts:
	module/Admin/src/Service/AplService.php

+++ module/Application/src/Service/AssemblyManager.php
@@ -631,7 +631,9 @@ class AssemblyManager
             
             $this->entityManager->detach($article);
             $this->entityManager->detach($producer);
-            $this->entityManager->detach($good);
+            if ($good){
+                $this->entityManager->detach($good);
+            }    
         }
         return;
     }
